simplify through explicit code
